Move RewindData and UnspentOutputs alongside DBreezeSingleThreadSession
