+++ Stratis.Bitcoin.Features.Consensus/CoinViews/FetchCoinsResponse.cs
@@ -1,4 +1,5 @@
 ﻿using NBitcoin;
+using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.Features.Consensus.CoinViews
 {
