+++ Stratis.Bitcoin.Features.Consensus/StakeValidator.cs
@@ -4,6 +4,7 @@ using NBitcoin.BouncyCastle.Math;
 using NBitcoin.Crypto;
 using NLog.Extensions.Logging;
 using Stratis.Bitcoin.Features.Consensus.CoinViews;
+using Stratis.Bitcoin.Utilities;
 using System;
 using System.Collections.Generic;
 using System.IO;
