+++ Stratis.Bitcoin.IntegrationTests/CoinViewTests.cs
@@ -7,6 +7,7 @@ using Stratis.Bitcoin.Base;
 using Stratis.Bitcoin.Base.Deployments;
 using Stratis.Bitcoin.Features.Consensus;
 using Stratis.Bitcoin.Features.Consensus.CoinViews;
+using Stratis.Bitcoin.Utilities;
 using System;
 using System.Globalization;
 using System.IO;
